+++ nextjs-metrics-viewer/interfaces/index.ts
@@ -1,0 +1,10 @@
+export interface formattedSeries {
+  name: string;
+  data: { x: string; y: string }[];
+}
+
+export enum Headers {
+  Date = "Date",
+  Category = "Category",
+  Value = "Value",
+}
